pyproject.toml
[project]
name = "google-maps-mcp"
version = "0.1.0"
description = "MCP server for Google Maps search capabilities"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "openai-agents>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "respx>=0.23.1",
]

